+++ data-transfer/src/main/java/it/eng/datatransfer/model/TransferStartMessage.java
@@ -3,6 +3,10 @@ package it.eng.datatransfer.model;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSetter;
@@ -47,12 +51,20 @@ import lombok.NoArgsConstructor;
 @Getter
 @JsonDeserialize(builder = TransferStartMessage.Builder.class)
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
+@Document(collection = "transfer_start_messages")
 public class TransferStartMessage extends AbstractTransferMessage {
 
+    @JsonIgnore
+    @JsonProperty(DSpaceConstants.ID)
+    @Id
+    private String id;
+    
 	@NotNull
 	@JsonProperty(DSpaceConstants.DSPACE_PROVIDER_PID)
 	private String providerPid;
 	
+	//	The dataAddress is only provided if the current transfer is a pull transfer 
+	//	and contains a transport-specific endpoint address for obtaining the data.
 	@JsonProperty(DSpaceConstants.DSPACE_DATA_ADDRESS)
 	private DataAddress dataAddress;
 	
@@ -70,6 +82,11 @@ public class TransferStartMessage extends AbstractTransferMessage {
 			return new Builder();
 		}
 		
+		public Builder id(String id) {
+        	message.id = id;
+        	return this;
+        }
+
 		@JsonSetter(DSpaceConstants.DSPACE_CONSUMER_PID)
 		public Builder consumerPid(String consumerPid) {
 			message.consumerPid = consumerPid;
@@ -89,6 +106,9 @@ public class TransferStartMessage extends AbstractTransferMessage {
 		}
 		
 		public TransferStartMessage build() {
+			if (message.id == null) {
+	               message.id = message.createNewId();
+	        }
 			Set<ConstraintViolation<TransferStartMessage>> violations 
 				= Validation.buildDefaultValidatorFactory().getValidator().validate(message);
 			if(violations.isEmpty()) {