+++ data-transfer/src/test/java/it/eng/datatransfer/model/ModelUtil.java
@@ -16,4 +16,10 @@ public class ModelUtil {
 	
 	public static final String TARGET = "urn:uuid:TARGET";
 	
+	public static TransferProcess TRANSFER_PROCESS_REQUESTED = TransferProcess.Builder.newInstance()
+			.agreementId(AGREEMENT_ID)
+			.consumerPid(CONSUMER_PID)
+			.providerPid(PROVIDER_PID)
+			.state(TransferState.REQUESTED)
+			.build();
 }