+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2023-08-03
+## [0.0.1] - 2024-06-27
+
+### Changed
+
+ - updated verified and finalized states
+ - separated consumer and provider callback addresses
+
+## [0.0.1] - 2024-06-21
 
 ### Added
 
+ - dataTransfer module, POC for REST pull artifact
+ - duplicate TransferProcess and ContractNegotiation with new status
+ - DataService, update method
+ - DataTransfer exception advice
+ 
 ### Changed
 
 ### Removed